[ 004 ] Layout fine-tuning
Layout of the webpage was changed: the whole content of the page is inside the 'content' div, including the scrolling behaviour. The size of the main panel is locked to specific values. There's now an area for pre-defined variables in style.css. For now I should focus on responsive web design to finalize the homepage and then I'll start introducing missing pages (most likely with the help of SSG tools).